Add R-kiosk
This was downloaded from addons.mozilla.org. It is in the public domain:

https://addons.mozilla.org/en-US/firefox/addon/r-kiosk/eula/

R-Kiosk extension is in public domain. It may be freely reproduced,
distributed, transmitted, used, modified, built upon, or otherwise used by
anyone for commercial or non-commercial purposes.

R-Kiosk extension is supplied as is. You use this extension at your own risk.