Set touchscreen to correct resolution
The chalkelec touchscreen has an odd native resolution (1280x800).  For some
reason this isn't automatically detected.  Manually set a modeline using
xrandr, and then reconfigure the touchscreen input to cover the whole screen.

This takes place in a .desktop file, so that it gets run late enough in the
process to have a permanent effect.